pyproject.toml
[project]
name = "alpha-miner"
version = "0.1.0"
description = "A股量化因子挖掘系统 — 行为金融学驱动"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "akshare>=1.10",
    "scipy>=1.11",
    "snownlp>=0.12",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "anthropic>=0.30",
    "baostock>=0.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live: marks tests that hit real APIs (deselect with '-m \"not live\"')",
]

[tool.hatch.build.targets.wheel]
packages = ["src", "cli"]
